Moved cache mechanism in auxiliary file
